Update script.js to remove console.logs.
To clarify, I only did those for debugging purposes. I know fully well the risks of logging info into the console that is sensitive: the only reason I committed without second thought is because 1- the chances of someone actually looking at this insignificant project's source code is infinitesimally small and 2- the STM API is completely free and only has a rate limit usage of around 40 calls per minute, if i recall correctly (after which is blocks the API endpoint for the month). 

If you're reading this, I also knew from the beginning that the refresh button is awfully implemented. calling the API every time the button is pressed would inevitably flood our available usages, but I have yet to find a way to cron job this process so users would no longer be able to manually refresh. That said, I'm just a humble student